Figured out that instead of 67 blocks, I actually need 125. I also optimized the rotation_init() a little bit and created less room for future mistakes by eliminating half the literal assignments. Also started trying to figure out a good way to setup the facelist pspimage, as the current version just uses shadow panes. After that part's done, it will only use sprites along the edges and corners, not the faces, which should reduce the growth rate of sprites significantly, and also lend itself to even further optimization based on list deltas which should allow me to do the fast selection routines like "select all" for very large grid sizes (100+). As it is, it gets slowed down on my haus machine around even 30x30x30 because it's horribly inefficient.
Encountered some problems with the new command.bmx not resizing the grid correctly, still looking into that - but I think it's still worth the effort to separate the commands from the controls in a modular way, so I'm going to continue stripping the meat away from the keyboard_input() routine in controller and putting it out to command.

That's all for now.

git-svn-id: http://isoblox.googlecode.com/svn@11 d6d271a0-1f4c-0410-9b3e-b771dbd1382d