Updated a few random things decided to hack together an inefficient mouse hovering thing, so I can continue with other things while I explore the most efficient algorithm to map screen coordinates to iso-coordinates.
Regarding the mouse control I understand all that's required now, it's just a matter of providing discrete enumerations to a couple of different sets of things.

I'm excited because eventually I'll be able to put a context-GUI and hover graphics when the mouse hovers over a block, or a selected area. Such as rotation. That will be much more intuitive.

git-svn-id: http://isoblox.googlecode.com/svn@63 d6d271a0-1f4c-0410-9b3e-b771dbd1382d